+++ app/models/comment.rb
@@ -1,9 +1,7 @@
 class Comment < ActiveRecord::Base
   include Concerns::Parentable
   include Concerns::Locale
-
   include Concerns::Spammable
-  spammable author_email: :email, content: :message
 
   STATUSES = %w(pending_moderation published)