+++ changelog.md
@@ -1,6 +1,6 @@
 1.0.1 - xxxx-xx-xx
 ------------------
-* Get rid of some dependencies that are currently not used.
+* Get rid of some dependencies that are currently not used: rakismet.
 
 
 1.0.0 - 2016-06-05